rest-api/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
django-cors-headers==3.5.0
djangorestframework-simplejwt==4.6.0